+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://book.async.rs/overview
 
 ## [Unreleased]
 
+Same as 6.5.0 with one change:
+
+`Config::max_connections_per_host()` is now properly named `Config::set_max_connections_per_host()`.
+
 ## [6.5.0]
+
+(Yanked)
 
 ### Added
 - `Config` has been stabilized and is now available by default!
@@ -15,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://book.async.rs/overview
 - `Config::max_connections_per_host` (Supported on `h1_client` and `curl_client`.)
 
 ### Deprecated
-- `H1Client::with_max_connections()` was been superseded by `Config::max_connections_per_host`.
+- `H1Client::with_max_connections()` will be superseded by `Config::max_connections_per_host`.
 
 ## [6.4.1] - 2021-05-19